Ban float and new compile options to potentially catch bugs

+++ src/Geometry/Line.h
@@ -5,18 +5,18 @@
 class Line {
  public:
   Line() {}
-  Line(const Vec<float>& p1, const Vec<float>& p2) : p1_(p1), p2_(p2) {}
+  Line(const Vec<double>& p1, const Vec<double>& p2) : p1_(p1), p2_(p2) {}
 
-  const Vec<float>& p1() const { return p1_; }
-  const Vec<float>& p2() const { return p2_; }
+  const Vec<double>& p1() const { return p1_; }
+  const Vec<double>& p2() const { return p2_; }
 
-  Vec<float> p1_;
-  Vec<float> p2_;
+  Vec<double> p1_;
+  Vec<double> p2_;
 
-  Vec<float> intercect(const Line& other) const;
+  Vec<double> intercect(const Line& other) const;
   Line perpendicular_bisector() const;
   bool orthogonal(const Line& other) const;
-  Vec<float> midpoint() const;
+  Vec<double> midpoint() const;
 
   void Print() const;
 };
